+++ src/efd/model/AssetTree.java
@@ -22,11 +22,11 @@ public class AssetTree extends Asset{
 	@Column(name = "NumberOwned", nullable=false )
 	@Required
 	@NotNull
-	private Integer numberOwned;
+	private Double numberOwned;
 	
 	@Column(name = "PricePerUnit" ,precision=10, scale=2)
 	@Digits(integer=10,fraction=2)
-	private BigDecimal pricePerUnit;
+	private Double pricePerUnit;
 
 	public String getTreeTypeEnteredName() {
 		return treeTypeEnteredName;
@@ -36,21 +36,23 @@ public class AssetTree extends Asset{
 		this.treeTypeEnteredName = treeTypeEnteredName;
 	}
 
-	public Integer getNumberOwned() {
+	public Double getNumberOwned() {
 		return numberOwned;
 	}
 
-	public void setNumberOwned(Integer numberOwned) {
+	public void setNumberOwned(Double numberOwned) {
 		this.numberOwned = numberOwned;
 	}
 
-	public BigDecimal getPricePerUnit() {
+	public Double getPricePerUnit() {
 		return pricePerUnit;
 	}
 
-	public void setPricePerUnit(BigDecimal pricePerUnit) {
+	public void setPricePerUnit(Double pricePerUnit) {
 		this.pricePerUnit = pricePerUnit;
 	}
+
+
